fix exception if no output metrics on empty DataFrame
